chore(drec-ui) moved files to appropiate folder
