+++ .env.example
@@ -58,3 +58,5 @@ SMTP_LOGIN=
 SMTP_PASSWORD=
 EMAIL_FROM_NAME=
 EMAIL_FROM=no-reply@example.com
+
+REGISTRATION_MESSAGE_TO_SIGN='I register as D-REC Origin user'
