+++ apps/drec-ui/src/apps/certificate/data/cached/cachedAllDevices.ts
