+++ apps/drec-ui/src/apps/certificate/data/cached/cachedFuelTypes.ts
