+++ apps/drec-ui/src/apps/certificate/data/index.ts
@@ -1,3 +1,4 @@
 export * from './getEnergyTypeImage';
-export * from './getMainFuelType';
-export * from './prepareBeneficiariesOptions';
+export * from './cached';
+export * from './handlers';
+export * from './fetching';
