--- apps/drec-ui/src/apps/certificate/data/prepareBeneficiariesOptions.ts
@@ -1,11 +1,0 @@
-import { OrganizationDTO } from '@energyweb/origin-drec-api-client';
-
-export const prepareBeneficiariesOptions = (beneficiaries: OrganizationDTO[]) => {
-    return beneficiaries?.map((beneficiary) => ({
-        label: `
-    ID: ${beneficiary.id},
-    Name: ${beneficiary.name}
-    `,
-        value: beneficiary.id
-    }));
-};
