+++ apps/drec-ui/src/apps/certificate/logic/getMainFuelType.ts
