+++ apps/drec-ui/src/apps/certificate/logic/index.ts
@@ -1,6 +1,4 @@
 export * from './blockchain-inbox';
 export * from './permissions';
-export * from './cached';
-export * from './handlers';
-export * from './fetching';
+export * from './getMainFuelType';
 export * from './menu';
