+++ apps/drec-ui/src/apps/certificate/logic/types.ts
@@ -1,6 +1,7 @@
 import { CertificateDTO, CertificateEvent } from '@energyweb/origin-drec-api-client';
+import { AxiosResponse } from 'axios';
 
-export type ApiDownloadFunction = (id: string) => Promise<any>;
+export type ApiDownloadFunction = (id: string) => Promise<AxiosResponse<any>>;
 
 export type DetailedCertificate = {
     blockchainPart: CertificateDTO;
