+++ apps/drec-ui/src/apps/device/view/pages/AllDevicesPage/AllDevicesPage.tsx
@@ -13,7 +13,7 @@ export const AllDevicesPage: FC = () => {
         return <CircularProgress />;
     }
 
-    if (allDevices?.length === 0) {
+    if (!isLoading && allDevices?.length === 0) {
         return <NoDevicesOwnedCard />;
     }
